backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.1.0
google-generativeai==0.8.4
faiss-cpu==1.9.0.post1
numpy==2.2.1
sse-starlette==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-auth==2.37.0
